BookInfo imports renamed to BookTracker
